3dhumanpose_main/requirements.txt
numpy~=1.19.2
Pillow~=8.2.0
torch~=1.8.0
matplotlib~=3.4.1
commentjson~=0.9.0
torchvision~=0.9.0
torchmetrics~=0.2.0
opencv-python==4.5.1.48